ai-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.34.0
pydantic>=2.10.0
pillow>=11.0.0
numpy>=2.0.0
python-multipart>=0.0.20
scikit-image>=0.25.0
pytest>=8.0.0
transformers>=4.40.0
torch>=2.0.0
paddleocr>=2.8.0
nudenet>=3.4.2
